타입 변경

+++ src/main/java/github/gunkim/climbingcalendar/infrastructure/jpa/climbinggym/entity/LevelEntity.java
@@ -12,8 +12,8 @@ import java.util.Objects;
 public class LevelEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private Integer climbingGymId;
+    private Long id;
+    private Long climbingGymId;
     private String color;
     private int startGrade;
     private int endGrade;
